Add config.toml option to disable loading rustc debug symbols on Windows
On Windows, calling `std::backtrace::Backtrace::capture()` will result
in `dbghelp.dll` being loaded and, when capturing a backtrace inside of
rustc, it will attempt to load the path of the PDB mentioned in
librustc_driver.dll.

This path will never exist on user machines because it only exists in
the environment where the toolchain was built. For some build systems
like BuildXL, attempting to load this path is detected and flagged by
the build system because it is outside of the project being built's
path.

This patch adds an option to `config.toml` which allows turning off this
functionality by setting the `SYMOPT_EX_NEVERLOADSYMBOLS` extended option.
Users who need a compiler with this behavior can then enable it when
building rustc.